fix: correct relative include paths in reorganized tests
