+++ tests/graph_traversals/test_topological_sort.c
@@ -26,7 +26,7 @@ int mock_printf(const char* format, ...)
 
 // Redirect printf to our mock
 #define printf mock_printf
-#include "../src/graph_traversals/topological_sort.c"
+#include "../../src/graph_traversals/topological_sort.c"
 #undef printf
 
 void test_topological_sort_dag()
